+++ kogito-quarkus-examples/process-business-calendar-example/pom.xml
@@ -1,3 +1,24 @@
+<?xml version="1.0"?>
+<!--
+
+    Licensed to the Apache Software Foundation (ASF) under one
+    or more contributor license agreements.  See the NOTICE file
+    distributed with this work for additional information
+    regarding copyright ownership.  The ASF licenses this file
+    to you under the Apache License, Version 2.0 (the
+    "License"); you may not use this file except in compliance
+    with the License.  You may obtain a copy of the License at
+
+      http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing,
+    software distributed under the License is distributed on an
+    "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+    KIND, either express or implied.  See the License for the
+    specific language governing permissions and limitations
+    under the License.
+
+-->
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
@@ -9,8 +30,8 @@
         <version>999-SNAPSHOT</version>
     </parent>
     
-    <artifactId>business-calendar-example</artifactId>
-    <name>Kogito Example :: Business Calendar</name>
+    <artifactId>process-business-calendar-example</artifactId>
+    <name>Kogito Example :: Process Business Calendar</name>
     
     <properties>
         <quarkus-plugin.version>3.8.6</quarkus-plugin.version>
