+++ kogito-quarkus-examples/process-business-calendar-example/src/main/docker/Dockerfile.native
