+++ kogito-quarkus-examples/process-business-calendar-example/src/main/resources/calendar.properties
