+++ kogito-quarkus-examples/process-business-calendar-example/src/main/resources/org/kie/calendar/transaction/BPMN2-BusinessCalendarBankTransaction.bpmn2
